+++ part2/server/user_microservice/app/models.py
@@ -11,6 +11,6 @@ class User(models.Model):
     email = models.CharField(max_length=200)
     name = models.CharField(max_length=200)
     type = models.IntegerField()
-    coins = models.FloatField()
+    coins = models.DecimalField(max_digits=10, decimal_places=2)
     def __str__(self):
         return self.name
